Add Test Harness and Refactor Matchmaking Logic
