+++ Destiny.ScrimTracker.Tests/ServiceTests/MatchMakingServiceTests.cs
@@ -1,0 +1,10 @@
+using NUnit.Framework;
+
+namespace Destiny.ScrimTracker.Tests.ServiceTests
+{
+    [TestFixture]
+    public class MatchMakingServiceTests
+    {
+        
+    }
+}
